Adding Functionality to the Search by Pokemon Name Field.
Similarily, using a for loop, I added functionality to loop through all of the heading tags which store the names of the pokemon and match it to what the user types. Then I filter then by only displaying those that match the search name and also alert the user the contents of the whole list item.